Refactor mysql, pull GtidReplicationPartitions into separate module
